# backend-go/docs/swagger.yaml
basePath: /api
definitions:
...
  presentation.ClubResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      logoUrl:
        type: string
      maxMembers:
        type: integer
      memberCount:
        type: integer
      monthlyFeeCents:
        type: integer
      monthlyFeeEuros:
        type: number
      name:
        type: string
      ownerId:
        type: integer
      ownerName:
        type: string
      ownerSlug:
        type: string
      slug:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
  presentation.CreateClubPaymentRequest:
    properties:
      amount_cents:
        type: integer
      customer_id:
        type: string
      membership_id:
        type: integer
      user_id:
        type: integer
    required:
    - amount_cents
    - customer_id
    - membership_id
    - user_id
    type: object
  presentation.CreateClubRequest:
    properties:
      description:
        type: string
      isActive:
        type: boolean
      logoUrl:
        type: string
      maxMembers:
        minimum: 1
        type: integer
      monthlyFeeCents:
        minimum: 0
        type: integer
      name:
        type: string
      ownerSlug:
        type: string
    required:
    - maxMembers
    - name
    type: object
...
  presentation.CreateUserRequest:
    properties:
      email:
        type: string
      fullName:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      roleId:
        type: integer
    required:
    - email
    - fullName
    - password
    type: object
...
  presentation.PaymentResponse:
    properties:
      amount_cents:
        type: integer
      amount_euros:
        type: number
      booking_id:
        type: integer
      class_enrollment_id:
        type: integer
      club_membership_id:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      id:
        type: integer
      payment_type:
        type: string
      provider:
        type: string
      status:
        type: string
      stripe_payment_intent_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  presentation.UpdateClubRequest:
    properties:
      description:
        type: string
      isActive:
        type: boolean
      logoUrl:
        type: string
      maxMembers:
        minimum: 1
        type: integer
      monthlyFeeCents:
        minimum: 0
        type: integer
      name:
        type: string
      ownerSlug:
        type: string
      status:
        type: string
    required:
    - maxMembers
    - name
    type: object
...
  presentation.UpdateUserRequest:
    properties:
      fullName:
        type: string
      isActive:
        type: boolean
      phone:
        type: string
      roleId:
        type: integer
    type: object
  presentation.UserResponse:
    properties:
      avatarUrl:
        type: string
      createdAt:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      phone:
        type: string
      roleId:
        type: integer
      roleName:
        type: string
      slug:
        type: string
      stripeCustomerId:
        type: string
      updatedAt:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API REST para gestión de instalaciones deportivas
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: PoliManage API
  version: "1.0"
paths:
...
  /api/payments/club:
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del pago de membresía
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/presentation.CreateClubPaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presentation.PaymentResponse'
      summary: Procesar pago de membresía
      tags:
      - payments
...
  /api/users:
    get:
      parameters:
      - description: Filtrar por rol
        in: query
        name: role_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presentation.UserResponse'
            type: array
      summary: Listar todos los usuarios
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del usuario
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/presentation.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presentation.UserResponse'
      summary: Crear un nuevo usuario
      tags:
      - users
  /api/users/{slug}:
    delete:
      parameters:
      - description: Slug del usuario
        in: path
        name: slug
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Eliminar un usuario
      tags:
      - users
    get:
      parameters:
      - description: Slug del usuario
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.UserResponse'
      summary: Obtener usuario por slug
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: Slug del usuario
        in: path
        name: slug
        required: true
        type: string
      - description: Datos a actualizar
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/presentation.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.UserResponse'
      summary: Actualizar un usuario
      tags:
      - users
  /clubs:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presentation.ClubResponse'
            type: array
      summary: Lista todos los clubs
      tags:
      - Clubs
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del club
        in: body
        name: club
        required: true
        schema:
          $ref: '#/definitions/presentation.CreateClubRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presentation.ClubResponse'
      summary: Crea un nuevo club
      tags:
      - Clubs
  /clubs/{slug}:
    delete:
      parameters:
      - description: Club slug
        in: path
        name: slug
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Elimina un club
      tags:
      - Clubs
    get:
      parameters:
      - description: Club slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.ClubResponse'
      summary: Obtiene un club por slug
      tags:
      - Clubs
    put:
      consumes:
      - application/json
      parameters:
      - description: Club slug
        in: path
        name: slug
        required: true
        type: string
      - description: Datos del club
        in: body
        name: club
        required: true
        schema:
          $ref: '#/definitions/presentation.UpdateClubRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.ClubResponse'
      summary: Actualiza un club existente
      tags:
      - Clubs
...
schemes:
- http
swagger: "2.0"